Moved update_obs_dataset and write_json_to_hdf5 to file_utils.py
